services/payments/requirements.txt
app==0.0.1
astroid==2.15.0
attrs==22.2.0
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
colorama==0.4.6
dateutil==2.8.2
dill==0.3.6
exceptiongroup==1.1.0
Flask==2.2.3
idna==3.4
importlib-metadata==6.0.0
iniconfig==2.0.0
isort==5.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
lazy-object-proxy==1.9.0
MarkupSafe==2.1.2
mccabe==0.7.0
packaging==23.0
platformdirs==3.1.1
pluggy==1.0.0
py-healthcheck==1.10.1
pylint==2.17.0
pytest==7.2.1
python-dotenv==1.0.0
requests==2.28.2
six==1.16.0
stripe==5.2.0
tomli==2.0.1
tomlkit==0.11.6
typing_extensions==4.5.0
urllib3==1.26.14
Werkzeug==2.2.3
wrapt==1.15.0
yapf==0.32.0
zipp==3.15.0
